Update mne/event.py
fix commit.

+++ mne/event.py
@@ -110,7 +110,7 @@ def define_target_events(events, reference_id, target_id, sfreq, tmin, tmax,
 
     new_events = []
     lag = []
-    for event in events.copy().astype('i8'):
+    for event in events.copy().astype('f8'):
         if event[2] == reference_id:
             lower = event[0] + imin
             upper = event[0] + imax
